+++ config/example_channelconfig.json
