+++ modules/translator.js
